[FEATURE] Add option to skip the check of a *.clone file

+++ recipes/typo3versions.rb
@@ -27,6 +27,7 @@
 base_directory       = node['typo3']['base_directory']
 git_new_workdir      = node['typo3']['path_git-new-workdir']
 shared_git_directory = node['typo3']['shared_git_directory'].gsub(/.git$/, '')
+check_clone_file     = node['typo3']['check_clone_file']
 
 #####################################
 # Maintain clones of TYPO3 versions
@@ -51,10 +52,33 @@ node['typo3']['install_branches'].each do |v|
   clone_file         = base_directory + "/" + branch + ".clone"
   marker_file        = base_directory + "/" + branch + ".cloned-by-chef"
 
-  # Only clone the repository if explicitely requested
-  if ::File.exists?(clone_file)
+  if ::File.exists?(destination)
 
-    if ::File.exists?(destination)
+    # Only reset if managed by Chef
+    if ::File.exists?(marker_file)
+      #####################################
+      # Update an existing directory
+      #####################################
+
+      # Update existing clone of TYPO3core (using git reset)
+      execute "Update TYPO3core for version #{branch}" do
+        cwd destination
+        umask 0022
+        command "git reset --hard #{reference}"
+      end
+
+    elsif check_clone_file and ::File.exists?(clone_file)
+      # New clone was requested, but the destination path exists already. Instead, clone into a temporary directory.
+      clone_into_temporary_directory = true
+
+    else
+      Chef::Log.debug("Skipping update of #{destination}: The repository is not managed by Chef.")
+    end
+  end
+
+  # Only clone the repository if explicitely requested
+  if not check_clone_file or ::File.exists?(clone_file)
+    if clone_into_temporary_directory
       # If the destination already exists, operate on a temporary folder and let the admin review the change
       # In order to use this folder, the admin needs to rename the temporary folder to the real destination name.
       prefix = branch + "_"
@@ -70,38 +94,26 @@ node['typo3']['install_branches'].each do |v|
       marker_file    = base_directory + "/" + tempname + ".cloned-by-chef"
     end
 
-    # Create initial clone of TYPO3core
-    # Use git-new-workdir share the main ".git" folder between all versions
-    execute "Create new TYPO3 working directory for version #{branch}" do
-      cwd base_directory
-      umask 0022
-      command <<-EOH
-        sh #{git_new_workdir} #{source} #{destination}
-        touch #{marker_file}
-        rm #{clone_file}
-      EOH
-    end
+    if not ::File.exists?(destination)
+      # Create initial clone of TYPO3core
+      # Use git-new-workdir share the main ".git" folder between all versions
+      execute "Create new TYPO3 working directory for version #{branch}" do
+        cwd base_directory
+        umask 0022
+        command <<-EOH
+          sh #{git_new_workdir} #{source} #{destination}
+          touch #{marker_file}
+          test -f #{clone_file} && rm #{clone_file} || true
+        EOH
+      end
 
-    execute "Create new local branch #{branch}" do
-      cwd destination
-      umask 0022
-      command "git checkout #{branch} || git checkout -b #{branch}; git reset --hard #{reference}"
-    end
-
-  elsif ::File.exists?(destination)
-
-    # Only reset if managed by Chef
-    if ::File.exists?(marker_file)
-
-      # Update existing clone of TYPO3core (using git reset)
-      execute "Update TYPO3core for version #{branch}" do
+      execute "Create new local branch #{branch}" do
         cwd destination
         umask 0022
-        command "git reset --hard #{reference}"
+        command "git checkout #{branch} || git checkout -b #{branch}; git reset --hard #{reference}"
       end
-    else
-      Chef::Log.debug("Skipping update of #{destination}: The repository is not managed by Chef.")
     end
+
   else
     Chef::Log.debug("Skipping clone of #{destination}: Missing clone file.")
   end
